insuperadmin/requirements.txt
django-hvad==1.8.0
djangorestframework==3.5.2